37 - locking down delte item permissions

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -7,7 +7,7 @@ const { hasPermission } = require("../utils");
 const Mutations = {
   async createItem(parent, args, ctx, info) {
     if (!ctx.request.userId) {
-      throw new Error('You must be logged in to do that!');
+      throw new Error("You must be logged in to do that!");
     }
 
     const item = await ctx.db.mutation.createItem(
@@ -17,9 +17,9 @@ const Mutations = {
           // This is how we create a releationship between Item and User
           user: {
             connect: {
-              id: ctx.request.userId
-            }
-          }
+              id: ctx.request.userId,
+            },
+          },
         },
       },
       info,
@@ -51,10 +51,17 @@ const Mutations = {
       `{
       id
       title
+      user { id }
     }`,
     );
     // 2. Check if the own that item, or have the permissions
-    // TODO:
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission),
+    );
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that");
+    }
     // 3. Delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -154,28 +161,34 @@ const Mutations = {
   },
   async updatePermissions(parent, args, ctx, info) {
     if (!ctx.request.userId) {
-      throw new Error('You must be logged in to do that!');
+      throw new Error("You must be logged in to do that!");
     }
 
-    const currentUser = await ctx.db.query.user({
-      where: {
-        id: ctx.request.userId
-      }
-    }, info);
-
-    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
-
-    return ctx.db.mutation.updateUser({
-      data: {
-        permissions: {
-          set: args.permissions
-        }
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId,
+        },
       },
-      where: {
-        id: args.userId
-      }
-    }, info);
-  }
+      info,
+    );
+
+    hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
+
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions,
+          },
+        },
+        where: {
+          id: args.userId,
+        },
+      },
+      info,
+    );
+  },
 };
 
 module.exports = Mutations;
